refactor(posts): Move `logger` into `lib/`
