+++ packages/posts/words/searchWords.js
@@ -1,5 +1,5 @@
 import _ from "lodash";
-import logger from "../logger";
+import logger from "../lib/logger";
 import {initializeWordSources} from "./wordSources";
 
 const searchWords = wordSearchParams => {
